backend/requirements.txt
flask
mongoengine
flask-cors
flask-jwt-extended
werkzeug
pillow
dnspython
python-dotenv
certifi
